Moved models to the db folder
