+++ bot/.sequelizerc
@@ -2,7 +2,7 @@ const path = require('path');
 
 module.exports = {
   'config': path.resolve('src/db/config', 'config.json'),
-  'models-path': path.resolve('src', 'models'),
+  'models-path': path.resolve('src/db', 'models'),
   'seeders-path': path.resolve('src/db', 'seeders'),
   'migrations-path': path.resolve('src/db', 'migrations')
 };
