+++ bot/src/events/GuildBanAdd.js
@@ -1,5 +1,5 @@
 import { EmbedBuilder, Events, AuditLogEvent } from 'discord.js';
-import { guildSettings } from '../models/guild-settings.js';
+import { guildSettings } from '../db/models/guild-settings.js';
 
 export default {
     name: Events.GuildBanAdd,
