+++ bot/src/events/GuildCreate.js
@@ -1,6 +1,6 @@
 import { EmbedBuilder, Events } from 'discord.js';
 import { abEmbedColour } from '../storage/consts.js';
-import { guildSettings } from '../models/guild-settings.js';
+import { guildSettings } from '../db/models/guild-settings.js';
 
 export default {
     name: Events.GuildCreate,
