+++ bot/src/events/MessageBulkDelete.js
@@ -1,5 +1,5 @@
 import { EmbedBuilder, Events, AttachmentBuilder } from 'discord.js';
-import { guildSettings } from '../models/guild-settings.js';
+import { guildSettings } from '../db/models/guild-settings.js';
 
 export default {
     name: Events.MessageBulkDelete,
