+++ bot/src/events/MessageDelete.js
@@ -1,5 +1,5 @@
 import {AttachmentBuilder, EmbedBuilder, Events} from 'discord.js';
-import { guildSettings } from '../models/guild-settings.js';
+import { guildSettings } from '../db/models/guild-settings.js';
 
 export default {
     name: Events.MessageDelete,
